fix(KEN-987): an empty projects list is one value, not a new one
Three components read the registered projects as
`s.settings?.projects ?? []`. Before the settings read lands that
selector mints a fresh array on every snapshot, so React sees a store
that changed on every render and the tree re-renders until it gives up:
"Maximum update depth exceeded", beside the warning that a getSnapshot
result should be cached. The Subscribe dialog and the Customize page
both froze on mount, and the available-package tree did until KEN-1132
fixed that one site without a test.

`projectsOf` in the new `settings-projects.ts` is the one selector all
three take, and the empty answer is one shared array. The projects slice
moves there whole, which is the seam that keeps `settings.ts` under its
size class; its baseline row goes.

Two guards. `available-package.dom.test.tsx` mounts the page with
settings unread, which is the state the app first draws it in, and
`settings.test.ts` pins the selector's identity. Both run red on
`?? []` restored.

`useCatalog` was the suspected cause and is not: all three callers pass
a catalog held in the nav store, and the page settles with the effect
untouched.
