+++ changelog.d/fixed/KEN-987.md
@@ -1,0 +1,3 @@
+- The Subscribe dialog and the Customize page no longer freeze when opened
+  before the settings read lands: their project list was read in a way that
+  made the app re-render without ever settling.
